Added remaining interfaces. Updated REST folder structure
