+++ sdks/cpp/connections/REST/include/interface/ISocketReader.h
